fix(frontend): race condition between authentication and navigation
There was a race condition on login/register.

Here's the scenario:
- A user enters their credentials on `AuthForm.tsx` and submits them.
- It calls the login mutation to perform the request on `POST /login`.
- If the request is successful, the `onSuccess` method is executed.
- `onSuccess` executes `login` from `AuthProvider` **as a synchronous method**.
- It doesn't wait for the response and executes the `onSuccess` method of `AuthForm.tsx`
which redirects the user.
- On redirection, `ProtectedRoute` is called. The session request might still be pending
in which case `ProtectedRoute` consider the user has unauthenticated and redirects them
to the homepage.
- This cause a flickering effect where the user authenticates successfully but get redirection
to the homepage.

This race condition was fixed by addressing several small issues:
- Remove navigation from `AuthForm.tsx` and move it after session refetch completes.
- Make login and logout async and await them.
- Stop using isLoading from TanStack (which gives the status of the first load **only**)
and use isFetching instead (which indicates if there is any request in flight).
